requirements.txt
requests<2.32
urllib3<2.0
python-dotenv
